+++ 02_Solidity/BLOQUE 1 - Conceptos básicos/Propiedades de transacciones y bloques/keccak.sol
@@ -2,15 +2,16 @@
 pragma solidity >=0.4.4 <0.8.9;
 pragma experimental ABIEncoderV2;
 
-contract hash{
-    
-    function calcularHash(string memory _cadena) public pure returns(bytes32){
+contract hash {
+    function calcularHash(string memory _cadena) public pure returns (bytes32) {
         return keccak256(abi.encodePacked(_cadena));
-        
     }
-    
-    function calcularHash(string memory _cadena, uint _k, address _direccion) public pure returns(bytes32){
+
+    function calcularHash(
+        string memory _cadena,
+        uint256 _k,
+        address _direccion
+    ) public pure returns (bytes32) {
         return keccak256(abi.encodePacked(_cadena, _k, _direccion));
-        
     }
 }
